+++ src/components/formcomponents/text/text.tsx
@@ -94,6 +94,12 @@ export const Text = (props: Props): JSX.Element | null => {
     validate: (value: string): string | true | undefined =>
       scriptInjection({ value, resources, shouldValidate: !!validateScriptInjection }),
     shouldUnregister: true,
+    setValueAs: (value: string): string | undefined => {
+      if (value === undefined || value === null || value === '') {
+        return undefined;
+      }
+      return Array.isArray(value) ? value : value?.trim();
+    },
   };
 
   const { onChange, ...rest } = register(idWithLinkIdAndItemIndex, shouldValidate(item, pdf) ? validationRules : undefined);